Pull Game from Team
Make the Game model a first class collection, and reference the ID
inside of the schedule (within the Team). This allows us the ability to
re-use a game between the two teams that share it. The downside is that
win/loss and team records need to be determined dynamically. Also
update the URLs to move schedule off of the team.