WIP: large scale additions for sessions and user games watched
Clean up the session support to better login and retrieve who is logged
in. Add the beginnings of user's games watched, with front end support
for clicking the games that they've watched. Add in backbone route
filters to make sure they are logged in before hitting the next page(s).