Add specific buttons for edit and watched games
Add buttons to each game to specify that the game has been watched, and
to edit the game (not yet implemented).  Refactor the way games are
stored under the user to include the notes information (which would be
added in an edit), and completed flag.  Reference a game by the game ID
rather than the game time UTC.