Add notes to a game
Add the ability to set a note on a game.  This is done by flipping the
current view to a game-notes view, which has an input for the current
note.  An update from there triggers the set on the note within the
game stored under the user.  Also display if a note was set in the UI
by button color.  And don’t display the notes/watched button for an
archived game, instead display undo archived button (needs to be
implemented).