Move most of the logic in sorting games to server
In an attempt to improve the speed of the UI, move most of the logic in determining where the game is placed in the UI to the server.  In doing so, a new controller has been created which will return all the games for a certain user, and includes if the game was watched and any notes.  It also sorts the games by the game time UTC.  This allows the UI to make 1 API call, and simply display the result.

Still need to handle updating the games notes/watched status, but this first pass should determine if the changes done improves the UI performance.