Add logic to update games
With the logic rewrite to use view models instead of the multiple API
calls in the UI, we needed an ability to update games for
notes/complete.  Instead of using a PUT, use a PATCH so that Backbone
just sends the game information rather than all the games (from the
view model in the original GET request).  WIthin this API, perform the
necessary calls to update the existing user's sportsWatched array with
the data sent in.