Minify JS/CSS
Add grunt tasks for minification of JS/CSS.  This required using
absolute paths in require instead of the /assets to /public translation
done in express (at least for now).  So any place /assets existed was
changed to /public.  Also used relative paths in main.js and in loading
of templates with the text plugin.

New production version of the view was created to include minified
versions of the JS and CSS.  Almond was also included which is used for
require "define" stanzas.  Minified versions are checked in and created
using the grunt default task.