Update team schedules when teams are selected
When a user selects a team on the team page, update the schedules of
teams selected.  This makes sure that the games exist for each team
tracked by the user.

Doing this required pulling the common logic out of the schedule
controller, and into the team-schedule library.  Then both controllers
were able to invoke the same code to update schedules of teams.