Add MLB teams/configuration
Add all the images for the MLB teams (and MLB logo).  Add all the data
on the MLB teams in the configuration, which includes a necessary
"teamId".  This teamId was added to the Team model.

The teamId is needed to pull down the team schedule CSV, which is keyed
off an ID for each team.  No idea why, nor the logic behind any
ordering of this ID.  It was a very painful, manual, process to find
each one.