Add MLB game builder, support for MLB games
Add an MLB game builder, which builds MLB games using the schedule
information found for each team.  Add support for displaying the MLB
teams for a user, and adding the games to the user's game manager.

Update the style a bit to better display the MLB logo, and a team.