Pull schedule from ICS rather than CSV
To get an actual date/time for the game, standardized, instead of a
"sometimes in your time zone, sometimes in another", use the ICS
version of the schedule rather than the CSV.  This will provide us with
a standard time for the game specified in the GMT time zone.  Using the
time.Date object, we can convert this to the specified time zone (New
York), and then convert that to our local time zone when storing it.

Aside from that, the numbers and fields are pretty much the same, so
the original logic should still apply here.  What we get however, is
that the game times are correct for any time zone, AND they are correct
for daylight saving times (which really suck).